+ use is_like flag for check match + add JSON export for log + calc distance when open view profile in smchat + wait for set liked operation completed before show match buttons

+++ oakclubbuild/Profile.h
@@ -71,7 +71,7 @@
     
     int num_Liked;
     int num_Viewed;
-    int distance;
+    double distance;
     int active;
     
     NSString *s_lastMessage;
@@ -134,7 +134,7 @@
 @property int unread_message;
 @property int num_Liked;
 @property int num_Viewed;
-@property int distance;
+@property double distance;
 @property int active;
 @property (strong, nonatomic) NSString *s_lastMessage;
 @property (strong, nonatomic) NSString *s_lastMessage_time;
